Final presentaion touchs

+++ lib/models/upcoming_requests/ending_mechanic_service.dart
@@ -20,7 +20,7 @@ class EndingMechanicServiceResponseModel {
   });
 
   String status;
-  int fare;
+  double fare;
   TotalTimeForService totalTimeForService;
 
   factory EndingMechanicServiceResponseModel.fromJson(
